[Decomp][Spell] Cohesion split: extract SpellAuraTrailing handlers to SpellAuraTrailing.cpp
