+++ src/game/WorldHandlers/SpellAuras.cpp
@@ -3327,149 +3327,10 @@ void Aura::PeriodicDummyTick()
     }
 }
 
-/**
- * @brief Applies or removes prevention of fleeing on feared targets.
- *
- * @param apply True to prevent fleeing; false to restore it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandlePreventFleeing(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
 
-    Unit::AuraList const& fearAuras = GetTarget()->GetAurasByType(SPELL_AURA_MOD_FEAR);
-    if (!fearAuras.empty())
-    {
-        if (apply)
-        {
-            GetTarget()->SetFeared(false, fearAuras.front()->GetCasterGuid());
-        }
-        else
-        {
-            GetTarget()->SetFeared(true);
-        }
-    }
-}
 
-/**
- * @brief Calculates bonus absorb values for mana shield effects.
- *
- * @param apply True to apply the shield; false to remove it.
- * @param Real True when processing the real aura state change.
- */
-void Aura::HandleManaShield(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
 
-    // prevent double apply bonuses
-    if (apply && (GetTarget()->GetTypeId() != TYPEID_PLAYER || !((Player*)GetTarget())->GetSession()->PlayerLoading()))
-    {
-        if (Unit* caster = GetCaster())
-        {
-            float DoneActualBenefit = 0.0f;
-            switch (GetSpellProto()->SpellFamilyName)
-            {
-                case SPELLFAMILY_MAGE:
-                    if (GetSpellProto()->SpellFamilyFlags & UI64LIT(0x0000000000008000))
-                    {
-                        // Mana Shield
-                        // +50% from +spd bonus
-                        DoneActualBenefit = caster->SpellBaseDamageBonusDone(GetSpellSchoolMask(GetSpellProto())) * 0.5f;
-                        break;
-                    }
-                    break;
-                default:
-                    break;
-            }
 
-            DoneActualBenefit *= caster->CalculateLevelPenalty(GetSpellProto());
-
-            m_modifier.m_amount += (int32)DoneActualBenefit;
-        }
-    }
-}
-
-void Aura::HandleArenaPreparation(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    Unit* target = GetTarget();
-
-    target->ApplyModFlag(UNIT_FIELD_FLAGS, UNIT_FLAG_PREPARATION, apply);
-
-    if (apply)
-    {
-        // max regen powers at start preparation
-        target->SetHealth(target->GetMaxHealth());
-        target->SetPower(POWER_MANA, target->GetMaxPower(POWER_MANA));
-        target->SetPower(POWER_ENERGY, target->GetMaxPower(POWER_ENERGY));
-    }
-    else
-    {
-        // reset originally 0 powers at start/leave
-        target->SetPower(POWER_RAGE, 0);
-    }
-}
-
-void Aura::HandleAuraMirrorImage(bool apply, bool Real)
-{
-    if (!Real)
-    {
-        return;
-    }
-
-    // Target of aura should always be creature (ref Spell::CheckCast)
-    Creature* pCreature = (Creature*)GetTarget();
-
-    if (apply)
-    {
-        // Caster can be player or creature, the unit who pCreature will become an clone of.
-        Unit* caster = GetCaster();
-
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 0, caster->getRace());
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 1, caster->getClass());
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 2, caster->getGender());
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 3, caster->GetPowerType());
-
-        pCreature->SetFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_CLONED);
-
-        pCreature->SetDisplayId(caster->GetNativeDisplayId());
-    }
-    else
-    {
-        const CreatureInfo* cinfo = pCreature->GetCreatureInfo();
-        const CreatureModelInfo* minfo = sObjectMgr.GetCreatureModelInfo(pCreature->GetNativeDisplayId());
-
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 0, 0);
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 1, cinfo->UnitClass);
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 2, minfo->gender);
-        pCreature->SetByteValue(UNIT_FIELD_BYTES_0, 3, 0);
-
-        pCreature->RemoveFlag(UNIT_FIELD_FLAGS_2, UNIT_FLAG2_CLONED);
-
-        pCreature->SetDisplayId(pCreature->GetNativeDisplayId());
-    }
-}
-
-void Aura::HandleAuraSafeFall(bool Apply, bool Real)
-{
-    // implemented in WorldSession::HandleMovementOpcodes
-
-    // only special case
-    if (Apply && Real && GetId() == 32474 && GetTarget()->GetTypeId() == TYPEID_PLAYER)
-    {
-        ((Player*)GetTarget())->ActivateTaxiPathTo(506, GetId());
-    }
-}
 
 /**
  * @brief Checks whether this aura is the last remaining effect on its holder.
